# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: PitMaster AI - BBQ Planning & Monitoring API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cooks:
    get:
      operationId: listCooks
      tags: [cooks]
      summary: List all cook sessions
      parameters:
        - name: grillId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
            enum: [planned, active, completed, cancelled, null]
      responses:
        "200":
          description: List of cook sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Cook"
    post:
      operationId: createCook
      tags: [cooks]
      summary: Create a new cook session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCookBody"
      responses:
        "201":
          description: Created cook
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cook"
  /cooks/{id}:
    get:
      operationId: getCook
      tags: [cooks]
      summary: Get a cook session by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cook session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cook"
        "404":
          description: Not found
    patch:
      operationId: updateCook
      tags: [cooks]
      summary: Update a cook session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCookBody"
      responses:
        "200":
          description: Updated cook
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cook"
        "404":
          description: Not found
    delete:
      operationId: deleteCook
      tags: [cooks]
      summary: Delete a cook session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /ai/chat:
    post:
      operationId: aiChat
      tags: [ai]
      summary: Send a message to the AI assistant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiChatBody"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiChatResponse"
  /ai/predict:
    post:
      operationId: aiPredict
      tags: [ai]
      summary: Predict cook start and finish times based on grill and food history
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiPredictBody"
      responses:
        "200":
          description: Prediction result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiPredictResponse"
  /forum/posts:
    get:
      operationId: listForumPosts
      tags: [forum]
      summary: List forum posts
      parameters:
        - name: category
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of forum posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ForumPost"
    post:
      operationId: createForumPost
      tags: [forum]
      summary: Create a forum post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateForumPostBody"
      responses:
        "201":
          description: Created post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForumPost"
  /forum/posts/{id}:
    get:
      operationId: getForumPost
      tags: [forum]
      summary: Get a forum post with its comments
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Forum post with comments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForumPostDetail"
        "404":
          description: Not found
  /forum/posts/{id}/like:
    patch:
      operationId: likeForumPost
      tags: [forum]
      summary: Like a forum post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForumPost"
  /forum/posts/{id}/comments:
    post:
      operationId: createForumComment
      tags: [forum]
      summary: Add a comment to a forum post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateForumCommentBody"
      responses:
        "201":
          description: Created comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForumComment"
...
  /alerts:
    get:
      operationId: listAlerts
      tags: [alerts]
      summary: List temperature alerts
      responses:
        "200":
          description: List of alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
    post:
      operationId: createAlert
      tags: [alerts]
      summary: Create a temperature alert
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAlertBody"
      responses:
        "201":
          description: Created alert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
  /alerts/{id}:
    delete:
      operationId: deleteAlert
      tags: [alerts]
      summary: Delete an alert
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Cook:
      type: object
      properties:
        id:
          type: integer
        grillId:
          type: ["integer", "null"]
        grillName:
          type: ["string", "null"]
        foodType:
          type: string
        weightLbs:
          type: ["number", "null"]
        targetTempF:
          type: ["number", "null"]
        cookTempF:
          type: ["number", "null"]
        status:
          type: string
          enum: [planned, active, completed, cancelled]
        plannedStartAt:
          type: ["string", "null"]
          format: date-time
        actualStartAt:
          type: ["string", "null"]
          format: date-time
        plannedEndAt:
          type: ["string", "null"]
          format: date-time
        actualEndAt:
          type: ["string", "null"]
          format: date-time
        notes:
          type: ["string", "null"]
        rating:
          type: ["integer", "null"]
        recipeId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - grillId
        - grillName
        - foodType
        - weightLbs
        - targetTempF
        - cookTempF
        - status
        - plannedStartAt
        - actualStartAt
        - plannedEndAt
        - actualEndAt
        - notes
        - rating
        - recipeId
        - createdAt
        - updatedAt
    CreateCookBody:
      type: object
      properties:
        grillId:
          type: ["integer", "null"]
        foodType:
          type: string
        weightLbs:
          type: ["number", "null"]
        targetTempF:
          type: ["number", "null"]
        cookTempF:
          type: ["number", "null"]
        status:
          type: string
          enum: [planned, active, completed, cancelled]
        plannedStartAt:
          type: ["string", "null"]
          format: date-time
        plannedEndAt:
          type: ["string", "null"]
          format: date-time
        notes:
          type: ["string", "null"]
        recipeId:
          type: ["integer", "null"]
      required:
        - foodType
    UpdateCookBody:
      type: object
      properties:
        grillId:
          type: ["integer", "null"]
        foodType:
          type: ["string", "null"]
        weightLbs:
          type: ["number", "null"]
        targetTempF:
          type: ["number", "null"]
        cookTempF:
          type: ["number", "null"]
        status:
          type: ["string", "null"]
          enum: [planned, active, completed, cancelled, null]
        plannedStartAt:
          type: ["string", "null"]
          format: date-time
        actualStartAt:
          type: ["string", "null"]
          format: date-time
        plannedEndAt:
          type: ["string", "null"]
          format: date-time
        actualEndAt:
          type: ["string", "null"]
          format: date-time
        notes:
          type: ["string", "null"]
        rating:
          type: ["integer", "null"]
        recipeId:
          type: ["integer", "null"]
...
    AiChatBody:
      type: object
      properties:
        message:
          type: string
        context:
          type: ["string", "null"]
          description: Optional JSON context (active cook, selected grill, etc.)
      required:
        - message
    AiChatResponse:
      type: object
      properties:
        reply:
          type: string
        suggestions:
          type: array
          items:
            type: string
      required:
        - reply
        - suggestions
    AiPredictBody:
      type: object
      properties:
        grillId:
          type: ["integer", "null"]
        foodType:
          type: string
        weightLbs:
          type: ["number", "null"]
        cookTempF:
          type: ["number", "null"]
        targetTempF:
          type: ["number", "null"]
        desiredFinishAt:
          type: ["string", "null"]
          format: date-time
      required:
        - foodType
    AiPredictResponse:
      type: object
      properties:
        estimatedDurationMinutes:
          type: number
        suggestedStartAt:
          type: ["string", "null"]
          format: date-time
        estimatedFinishAt:
          type: ["string", "null"]
          format: date-time
        confidence:
          type: string
          enum: [low, medium, high]
        rationale:
          type: string
        tips:
          type: array
          items:
            type: string
      required:
        - estimatedDurationMinutes
        - suggestedStartAt
        - estimatedFinishAt
        - confidence
        - rationale
        - tips
    ForumPost:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        category:
          type: string
        authorName:
          type: string
        authorAvatar:
          type: ["string", "null"]
        likesCount:
          type: integer
        commentsCount:
          type: integer
        imageUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - content
        - category
        - authorName
        - authorAvatar
        - likesCount
        - commentsCount
        - imageUrl
        - createdAt
        - updatedAt
    ForumComment:
      type: object
      properties:
        id:
          type: integer
        postId:
          type: integer
        content:
          type: string
        authorName:
          type: string
        authorAvatar:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - postId
        - content
        - authorName
        - authorAvatar
        - createdAt
    ForumPostDetail:
      type: object
      properties:
        post:
          $ref: "#/components/schemas/ForumPost"
        comments:
          type: array
          items:
            $ref: "#/components/schemas/ForumComment"
      required:
        - post
        - comments
    CreateForumPostBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        category:
          type: string
        authorName:
          type: string
        authorAvatar:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
      required:
        - title
        - content
        - category
        - authorName
    CreateForumCommentBody:
      type: object
      properties:
        content:
          type: string
        authorName:
          type: string
        authorAvatar:
          type: ["string", "null"]
      required:
        - content
        - authorName
...
    Alert:
      type: object
      properties:
        id:
          type: integer
        cookId:
          type: ["integer", "null"]
        probeNumber:
          type: ["integer", "null"]
        alertType:
          type: string
          enum: [min_temp, max_temp, target_reached, stall_detected]
        thresholdTempF:
          type: number
        message:
          type: string
        isActive:
          type: boolean
        triggeredAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - cookId
        - probeNumber
        - alertType
        - thresholdTempF
        - message
        - isActive
        - triggeredAt
        - createdAt
    CreateAlertBody:
      type: object
      properties:
        cookId:
          type: ["integer", "null"]
        probeNumber:
          type: ["integer", "null"]
        alertType:
          type: string
          enum: [min_temp, max_temp, target_reached, stall_detected]
        thresholdTempF:
          type: number
        message:
          type: string
      required:
        - alertType
        - thresholdTempF
        - message
...