# lib/api-zod/src/generated/types/forumPost.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PitMaster AI - BBQ Planning & Monitoring API
 * OpenAPI spec version: 0.1.0
 */

export interface ForumPost {
  id: number;
  title: string;
  content: string;
  category: string;
  authorName: string;
  /** @nullable */
  authorAvatar: string | null;
  likesCount: number;
  commentsCount: number;
  /** @nullable */
  imageUrl: string | null;
  createdAt: Date;
  updatedAt: Date;
}